fix: add serve directory
