+++ src/workflow-parser.test.ts
@@ -13,7 +13,7 @@ import type { RawStep, Step, WorkflowServeOptions } from "./types";
 import { getRequest, WORKFLOW_ENDPOINT } from "./test-utils";
 import { formatWorkflowError, QStashWorkflowError } from "./error";
 import { Client } from "@upstash/qstash";
-import { processOptions } from "./serve";
+import { processOptions } from "./serve/options";
 
 describe("Workflow Parser", () => {
   describe("validateRequest", () => {
